requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.6
aiosignal==1.3.1
async-timeout==4.0.3
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.3.2
contourpy==1.3.0
cycler==0.12.1
filelock==3.16.1
fonttools==4.54.0
frozenlist==1.4.1
fsspec==2024.9.0
idna==3.10
Jinja2==3.1.4
kiwisolver==1.4.7
lightning==2.4.0
lightning-utilities==0.11.7
MarkupSafe==2.1.5
matplotlib==3.9.2
mpmath==1.3.0
multidict==6.1.0
nba_api==1.5.2
networkx==3.3
numpy==1.26.4
packaging==24.1
pandas==2.2.3
pillow==10.4.0
pyaml==24.7.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
pytorch-lightning==2.4.0
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
seaborn==0.13.2
six==1.16.0
sympy==1.13.3
torch==2.2.2
torchmetrics==1.4.2
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
yarl==1.12.1
